Moving Storage to its own package
