--- java/core/src/main/java/co/worklytics/psoxy/FileHandlerStrategy.java
@@ -1,5 +1,0 @@
-package co.worklytics.psoxy;
-
-public interface FileHandlerStrategy {
-    FileHandler get(String fileName);
-}
