+++ java/core/src/main/java/co/worklytics/psoxy/storage/FileHandlerStrategy.java
@@ -1,0 +1,7 @@
+package co.worklytics.psoxy.storage;
+
+import co.worklytics.psoxy.storage.FileHandler;
+
+public interface FileHandlerStrategy {
+    FileHandler get(String fileName);
+}
