+++ java/impl/cmd-line/src/main/java/co/worklytics/psoxy/CmdLineModule.java
@@ -2,13 +2,8 @@ package co.worklytics.psoxy;
 
 import co.worklytics.psoxy.gateway.ConfigService;
 import co.worklytics.psoxy.gateway.impl.EnvVarsConfigService;
-import co.worklytics.psoxy.impl.CSVFileHandler;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
-
-import javax.inject.Named;
 
 @Module
 public interface CmdLineModule {
