+++ java/impl/cmd-line/src/main/java/co/worklytics/psoxy/Handler.java
@@ -1,5 +1,6 @@
 package co.worklytics.psoxy;
 
+import co.worklytics.psoxy.storage.FileHandlerStrategy;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.api.client.util.Lists;
 import com.google.cloud.secretmanager.v1.AccessSecretVersionResponse;
@@ -10,7 +11,6 @@ import lombok.SneakyThrows;
 import javax.inject.Inject;
 import java.io.File;
 import java.io.FileReader;
-import java.util.Collections;
 
 //@NoArgsConstructor(onConstructor_ = @Inject) //q: compile complaints - lombok annotation processing not reliable??
 public class Handler {
